requirements.txt
Flask==2.1.2
SQLAlchemy==1.4.27
Flask-Cors==3.0.10
requests==2.26.0
pytesseract==0.3.8
Pillow==8.4.0
openai==0.11.3